Combine resource authorizations with All and Any

+++ src/Mde.WishList.Api.Application/TodoItems/Commands/DeleteTodoItem/DeleteTodoItemCommand.cs
@@ -1,5 +1,6 @@
 ﻿using Mde.WishList.Api.Application.Common.Exceptions;
 using Mde.WishList.Api.Application.Common.Interfaces;
+using Mde.WishList.Api.Application.Common.Security;
 using Mde.WishList.Api.Domain.Entities;
 using MediatR;
 using System.Threading;
@@ -7,6 +8,8 @@ using System.Threading.Tasks;
 
 namespace Mde.WishList.Api.Application.TodoItems.Commands.DeleteTodoItem
 {
+
+    [Authorize]
     public class DeleteTodoItemCommand : IRequest
     {
         public int Id { get; set; }
@@ -15,10 +18,12 @@ namespace Mde.WishList.Api.Application.TodoItems.Commands.DeleteTodoItem
     public class DeleteTodoItemCommandHandler : IRequestHandler<DeleteTodoItemCommand>
     {
         private readonly IApplicationDbContext _context;
+        private readonly IResourceAuthorizationService _resourceAuthorizationService;
 
-        public DeleteTodoItemCommandHandler(IApplicationDbContext context)
+        public DeleteTodoItemCommandHandler(IApplicationDbContext context, IResourceAuthorizationService resourceAuthorizationService)
         {
             _context = context;
+            _resourceAuthorizationService = resourceAuthorizationService;
         }
 
         public async Task<Unit> Handle(DeleteTodoItemCommand request, CancellationToken cancellationToken)
@@ -29,6 +34,12 @@ namespace Mde.WishList.Api.Application.TodoItems.Commands.DeleteTodoItem
             {
                 throw new NotFoundException(nameof(TodoItem), request.Id);
             }
+
+            if (!await _resourceAuthorizationService.AuthorizeAny(entity, Policies.MustBeCreator, Policies.MustBeAdmin))
+            {
+                throw new ForbiddenAccessException();
+            }
+
 
             _context.TodoItems.Remove(entity);
